refactor: updated the path using index.ts
